% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_NorthernPike}
\alias{data_NorthernPike}
\title{Capture-recapture experiment on Northern Pike in Mille Lacs, MN, in 2005.}
\format{
\subsection{\code{data_NorthernPike}}{

A data frame with many rows and 4 columns

\describe{
\item{\code{cap_hist}.}{Capture history (10, 01, or 11). Note that
\eqn{n_{10}=n_1-m_2}{n10=n1-m2}; \eqn{n_{01}=n_2-m_2}{n01=n2-m2}; and \eqn{n_{11}=m2}{n11=m2}}
\item{\code{freq}.}{Number of fish with this history. Always 1}
\item{\code{Sex}.}{Sex of the fish. M=Male; F=Female}
\item{\code{Length}}{Length of the fish in inches.}
}
}
}
\usage{
data(data_NorthernPike)
}
\description{
Fish were tagged on the spawning grounds and recovered in the summer gillnet assessment.
Fish were double tagged, and a tag loss analysis
showed that tag loss was negligible. It will be ignored here.
Length was measured a both times and didn't not change
very much between the two sampling occasions. The value
recorded below is the average of the two lengths if both
lengths were present.
Fish that were not sexed or measured for length are ignored and not included
}
\keyword{datasets}
