% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_spas_harrison}
\alias{data_spas_harrison}
\title{Estimating abundance of salmon - SPAS - Harrison River}
\format{
\subsection{\verb{data_spas harrison}}{

A data frame with many rows and 2 columns

\describe{
\item{\code{cap_hist}.}{Capture history of the form `week..area' for fish that are released on
week and recaptured
area ; '0..area' for unmarked fish newly captured in area; 'week..0' for fish
released in week  but never recaptured.}
\item{\code{freq}.}{Number of fish with this history.}
}
}
}
\usage{
data(data_spas_harrison)
}
\description{
Incoming sockeye salmon are captured on a first wheel, tagged with color tags
that vary by week, and recaptured on several spawning areas.
}
\keyword{datasets}
