% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_yukon_reverse}
\alias{data_yukon_reverse}
\title{Yukon River data used for Reverse Capture-Recapture example.}
\format{
\code{data_yukon_reverse}
A data frame with many rows and 4 columns

\describe{
\item{\code{cap_hist}.}{Capture history with possible histories as noted below}
\item{\code{freq}.}{Number of fish with this history.}
\item{\code{SE}.}{SE of the number of fish with this history}
}
}
\usage{
data(data_yukon_reverse)
}
\description{
Data from
Hamazaki, T. and DeCovich, N. (2014).
Application of the Genetic Mark–Recapture Technique for Run Size Estimation of Yukon River Chinook Salmon.
North American Journal of Fisheries Management, 34, 276-286.
DOI: 10.1080/02755947.2013.869283
This is the data from the 2011 data in Table 2 of the above paper.
Estimated that total escapement to Canada (plus  harvest)
was 66,225 (SE 1574)
Estimated that proportion of stock that was Canadian was .34644 (SE .030)
We converted this into a "sample size" and number of fish with Cdn genetics
that gave the same SE.
}
\keyword{datasets}
