% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_AICc.R
\name{LP_AICc}
\alias{LP_AICc}
\title{Create an AIC table comparing multiple LP fits}
\usage{
LP_AICc(...)
}
\arguments{
\item{...}{Series of LP fits}
}
\value{
An data frame with an AICc table and model weights etc
}
\description{
This will take a series of LP fits and computes the usual AICc table and model weights
}
\examples{

data(data_rodli)
mt <- Petersen::LP_fit(data=data_rodli, p_model=~..time)
m0 <- Petersen::LP_fit(data=data_rodli, p_model=~1)
Petersen::LP_AICc(m0,mt)

}
