% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_BTSPAS_NonDiag_fit.R
\name{LP_BTSPAS_NonDiag_fit}
\alias{LP_BTSPAS_NonDiag_fit}
\title{Wrapper (*_fit) to call the Time Stratified Petersen Estimator
with NON-Diagonal Entries function in BTSPAS.}
\usage{
LP_BTSPAS_NonDiag_fit(
  data,
  p_model = ~1,
  p_model_cov = NULL,
  jump.after = NULL,
  logitP.fixed = NULL,
  logitP.fixed.values = NULL,
  InitialSeed = ceiling(stats::runif(1, min = 0, max = 1e+06)),
  trace = FALSE
)
}
\arguments{
\item{data}{Data frame containing the variables:
\itemize{
\item \strong{cap_hist} Capture history (see details below)
\item \strong{freq} Number of times this capture history was observed
}
plus any other covariates (e.g. discrete strata and/or continuous covariates) to be used
in the model fitting.}

\item{p_model}{Model for the captured probabilities. This can reference
other variables in the data frame, plus a special reserved term \code{..time} to indicate
a time dependence in the capture probabilities. For example, \code{p_model=~1} would indicate
that the capture probabilities are equal across the sampling events;
\code{p_model=~..time} would indicate that the capture probabilities vary by sampling events;
\code{p_model=~sex*..time} would indicate that the capture probabilites vary across
all combination of sampling events (\code{..time}) and a stratification variable (\code{sex}). The \code{sex} variable
also needs to be in the data frame.

For some models (e.g., tag loss models), the \code{..time} variable cannot be used because
the models only have one capture probability (e.g., only for event 1).}

\item{p_model_cov}{Data frame with covariates for the model for prob capture at second sampling event. If this
data frame is given, it requires one line for each of the temporal strata at the second sampling event (even
if missing in the \code{data} that has the capture histories) with one variable being \code{..time}
to represent
the second temporal stratum.}

\item{jump.after}{A numeric vector with elements belonging to \code{time}.
In some cases, the spline fitting the population numbers should be allowed
to jump.  For example, the population size could take a jump when hatchery
released fish suddenly arrive at the trap.  The jumps occur AFTER the strata
listed in this argument.}

\item{logitP.fixed}{A numeric vector (could be null) of the time strata
where the logit(P) would be fixed. Typically, this is used when the capture
rates for some strata are 0 and logit(P) is set to -10 for these strata. The
fixed values are given in \code{logitP.fixed.values}}

\item{logitP.fixed.values}{A numerical vector (could be null) of the fixed
values for logit(P) at strata given by logitP.fixed. Typically this is used
when certain strata have a 0 capture rate and the fixed value is set to -10
which on the logit scale gives $p_i$ essentially 0. Don't specify values such
as -50 because numerical problems could occur in WinBugs/OpenBugs.}

\item{InitialSeed}{Numeric value used to initialize the random numbers used
in the MCMC iterations.}

\item{trace}{Internal tracing flag.}
}
\value{
An MCMC object with samples from the posterior distribution. A
series of graphs and text file are also created in the results along with
a summary object
}
\description{
Takes the data structure as described below, and uses Bayesian methods to fit a fit a spline
through the population numbers and a hierarchical model for the trap
efficiency over time.  An MCMC object
is also created with samples from the posterior.
}
\details{
Use the \code{Petersen::LP_BTSPAS_Diag_fit} function for cases
where recaptures take place in a single stratum (diagonal case).

The frequency variable (\code{freq} in the \code{data} argument) is the number of animals with the corresponding capture history.

Capture histories (\code{cap_hist} in the \code{data} argument) are character values of the format
\code{xx..yy} is a capture_history where \code{xx} and \code{yy} are the temporal stratum
(e.g., julian week) and \code{'..'} separates
the two temporal strata.
If a fish is released in temporal stratum and never captured again, then \code{yy} is set to 0;
if a fish is newly captured in temporal stratum \code{yy}, then \code{xx} is set to zero.
For example, a capture history of \code{23..23} indicates animals released in temporal stratum
23 and recaptured in temporal stratum 23; a capture history of \code{23..00}
indicates animals released in temporal stratum
23 and never seen again; a capture history of \code{00..23}
indicates animals newly captured in temporal stratum
23 at the second sampling event.

In the non-diagonal case, fish are allowed to move among temporal strata.

It is not necessary to label the temporal strata starting at 1; BTSPAS will treat the smallest
value of the temporal strata seen as the first stratum and will interpolate for temporal strata
without any data. Temporal strata labels should be numeric, i.e., do NOT use A, B, C etc.
}
