% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_est_adjust.R
\name{LP_est_adjust}
\alias{LP_est_adjust}
\title{Estimate abundance after empirical adjustments for various factors.}
\usage{
LP_est_adjust(
  N_hat,
  N_hat_SE,
  conf_level = 0.95,
  tag.retention.est = NULL,
  tag.retention.se = NULL,
  tag.reporting.est = NULL,
  tag.reporting.se = NULL,
  n.sim = 10000,
  trace = FALSE
)
}
\arguments{
\item{N_hat}{Estimate of N that will be adjusted}

\item{N_hat_SE}{SE of the N_hat}

\item{conf_level}{The expected coverage for confidence intervals on N.}

\item{tag.retention.est}{Estimated tag retention probability}

\item{tag.retention.se}{Estimated SE of tag retention probability}

\item{tag.reporting.est}{Estimated tag reporting probability}

\item{tag.reporting.se}{Estimated SE of tag reporting probability}

\item{n.sim}{Number of simulation runs to make}

\item{trace}{If trace flag is set in call when estimating functions}
}
\value{
An list object with a summary data frame and a data frame with the adjustment factors.
}
\description{
This will take a previous fit and return estimates of abundance after making various empirical adjustments
}
\details{
The estimate and SE are converted to a beta distribution with equivalent
mean and SD as the estimate and se, and then simulated a large number of times
to get the mean and sd of all adjustments applied together.
Then the abundance is simulated (on the log scale), the product taken, and
the mean, sd, ci estimated directly.
}
\examples{

data(data_rodli)
rodli.fit <- Petersen::LP_fit(data=data_rodli, p_model=~..time)
rodli.est <- Petersen::LP_est(rodli.fit)
Petersen::LP_est_adjust(rodli.est$summary$N_hat, rodli.est$summary$N_hat_SE,
          tag.retention.est=.90, tag.retention.se=.05)
}
\author{
Schwarz, C. J. \email{cschwarz.stat.sfu.ca@gmail.com}.
}
