% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_IS_print.R
\name{LP_IS_print}
\alias{LP_IS_print}
\title{Print the results from a fit a Lincoln-Petersen Model with incomplete stratification}
\usage{
LP_IS_print(IS.results)
}
\arguments{
\item{IS.results}{Results from fitting an incomplete stratification model}
}
\description{
Print the results from a fit a Lincoln-Petersen Model with incomplete stratification
}
\examples{

data(data_wae_is_short)
res <- Petersen::LP_IS_fit(data=data_wae_is_short, p_model=~-1 + ..cat:..time)
LP_IS_print(res)

}
