% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_SPAS_est.R
\name{LP_SPAS_est}
\alias{LP_SPAS_est}
\title{Extract estimates of abundance after SPAS fit}
\usage{
LP_SPAS_est(LP_SPAS_fit, conf_level = 0.95, trace = FALSE)
}
\arguments{
\item{LP_SPAS_fit}{A result of an call to fitting at SPAS object.}

\item{conf_level}{The expected coverage for confidence intervals on N.}

\item{trace}{If trace flag is set in call when estimating functions}
}
\value{
An list object with abundance estimates
}
\description{
This will take a previous fit and return estimates of abundance.
}
\examples{

#data(data_rodli)
#rodli.fit <- Petersen::LP_fit(data=data_rodli, p_model=~..time)
#Petersen::LP_est(rodli.fit, N_hat=~1)
}
\author{
Schwarz, C. J. \email{cschwarz.stat.sfu.ca@gmail.com}.
}
