% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_SPAS_fit.R
\name{LP_SPAS_fit}
\alias{LP_SPAS_fit}
\title{Fit a Stratified-Petersen SPAS modelB.}
\usage{
LP_SPAS_fit(
  data,
  model.id = "Base model",
  autopool = FALSE,
  row.pool.in = NULL,
  col.pool.in = NULL,
  min.released = 100,
  min.inspected = 50,
  min.recaps = 50,
  min.rows = 1,
  min.cols = 1
)
}
\arguments{
\item{data}{Data frame containing the variables:
\itemize{
\item \strong{cap_hist} Capture history (see details below)
\item \strong{freq} Number of times this capture history was observed
}
plus any other covariates (e.g. discrete strata and/or continuous covariates) to be used
in the model fitting.}

\item{model.id}{Character string identifying the name of the model.}

\item{autopool}{Should the automatic pooling algorithms be used.
Give more details here on these rule work.}

\item{row.pool.in, col.pool.in}{Vectors (character/numeric) of length s and t respectively. These identify the rows/columns to be pooled before the analysis is done.
The vectors consists of entries where pooling takes place if the entries are the same. For example, if s=4, then
row.pool.in = c(1,2,3,4) implies no pooling because all entries are distinct; row.pool.in=c("a","a","b","b") implies that the
first two rows will be pooled and the last two rows will be pooled. It is not necessary that row/columns be continuous to be pooled, but
this is seldom sensible. A careful choice of pooling labels helps to remember what as done, e.g. row.pool.in=c("123","123","123","4") indicates
that the first 3 rows are pooled and the 4th row is not pooled. Character entries ensure that the resulting matrix is sorted properly (e.g. if
row.pool.in=c(123,123,123,4), then the same pooling is done, but the matrix rows are sorted rather strangely.}

\item{min.released}{Minimum number of releases in a pooled row}

\item{min.inspected}{Miniumum number of inspections in a pooled column}

\item{min.recaps}{Minimum number of recaptures before any rows can be pooled}

\item{min.rows, min.cols}{Minimum number or rows and columns after pooling}
}
\value{
A list with many entries including the fit from SPAS and a summary line
in standard format. Refer to the vignettes for more details.
}
\description{
This function is a wrapper to fits a SPAS model. Consult the SPAS package for more details.
}
\examples{
data(data_spas_harrison)

mod1 <- Petersen::LP_SPAS_fit(data=data_spas_harrison,
                      model.id="Pooling rows 5/6",
                      row.pool.in=c(1,2,3,4,56,56),
                      col.pool.in=c(1,2,3,4,5,6))
SPAS::SPAS.print.model(mod1$fit)
mod1$summary
}
